Release 0.2.0 (#431)

+++ test/decomposer/test_xzx_decomposer.py
@@ -4,10 +4,10 @@ import math
 
 import pytest
 
-from opensquirrel.decomposer.aba_decomposer import XZXDecomposer
-from opensquirrel.decomposer.general_decomposer import check_gate_replacement
-from opensquirrel.default_gates import CNOT, CR, H, I, Rx, Rz, S, X, Z
-from opensquirrel.ir import BlochSphereRotation, Float, Gate
+from opensquirrel import CNOT, CR, H, I, Rx, Rz, S, X, Z
+from opensquirrel.ir import BlochSphereRotation, Gate
+from opensquirrel.passes.decomposer import XZXDecomposer
+from opensquirrel.passes.decomposer.general_decomposer import check_gate_replacement
 
 
 @pytest.fixture(name="decomposer")
@@ -25,16 +25,16 @@ def test_identity(decomposer: XZXDecomposer) -> None:
     ("gate", "expected_result"),
     [
         (CNOT(0, 1), [CNOT(0, 1)]),
-        (CR(2, 3, Float(2.123)), [CR(2, 3, Float(2.123))]),
-        (S(0), [Rz(0, Float(math.pi / 2))]),
-        (Z(0), [Rz(0, Float(math.pi))]),
-        (Rz(0, Float(0.9)), [Rz(0, Float(0.9))]),
-        (X(0), [Rx(0, Float(math.pi))]),
-        (Rx(0, Float(0.123)), [Rx(0, Float(0.123))]),
-        (H(0), [Rx(0, Float(math.pi / 2)), Rz(0, Float(math.pi / 2)), Rx(0, Float(math.pi / 2))]),
+        (CR(2, 3, 2.123), [CR(2, 3, 2.123)]),
+        (S(0), [Rz(0, math.pi / 2)]),
+        (Z(0), [Rz(0, math.pi)]),
+        (Rz(0, 0.9), [Rz(0, 0.9)]),
+        (X(0), [Rx(0, math.pi)]),
+        (Rx(0, 0.123), [Rx(0, 0.123)]),
+        (H(0), [Rx(0, math.pi / 2), Rz(0, math.pi / 2), Rx(0, math.pi / 2)]),
         (
             BlochSphereRotation(qubit=0, angle=5.21, axis=(1, 2, 3), phase=0.324),
-            [Rx(0, Float(0.43035280630630446)), Rz(0, Float(-1.030183660156084)), Rx(0, Float(-0.7456524007888308))],
+            [Rx(0, 0.43035280630630446), Rz(0, -1.030183660156084), Rx(0, -0.7456524007888308)],
         ),
     ],
     ids=["CNOT", "CR", "S", "Y", "Ry", "X", "Rx", "H", "arbitrary"],
